tribunal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tribunal"
version = "0.3.0"
description = "Enterprise-grade discipline for Claude Code — TDD enforcement, quality gates, and team standards."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "thebotclub" }]
keywords = ["claude", "claude-code", "tdd", "quality", "hooks", "ai", "coding-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://tribunal.dev"
Repository = "https://github.com/thebotclub/tribunal"
Documentation = "https://github.com/thebotclub/tribunal/blob/main/README.md"

[project.scripts]
tribunal = "tribunal.cli:main"
tribunal-gate = "tribunal.gate:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tribunal"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.sdist]
include = ["src/tribunal"]
